Fixed temporal query
Temporal queries to not play nice with SF objects. Because of that, the SURFMod files have records still dating all the way back to 1975.

By reordering operations so that temporal query occurs before transformation to sf object (as was done for CEDENMod data), the temporal query was successful. Data is now restricted to project date range.

Updated SURFMod files are in the Output folder.